変更！！！

+++ to-do-app-sample/src/main/kotlin/com/example/domain/todo/delete/DeleteToDoInteractor.kt
@@ -1,6 +1,8 @@
 package com.example.domain.todo.delete
 
+import com.example.domain.todo.AuthorId
 import com.example.domain.todo.ToDoRepository
+import com.example.domain.todo.Todo
 
 class DeleteToDoInteractor(
     private val toDoRepository: ToDoRepository
@@ -9,7 +11,15 @@ class DeleteToDoInteractor(
         toDoRepository
             .find(toDoRepository.idFromString(request.id))
             .onSuccess {
-                toDoRepository.delete(it)
+                if (it.canBeDeleted(AuthorId(request.memberId))) {
+                    toDoRepository.delete(it)
+                }
             }
+    }
+}
+
+data class OperatorPrivilege(val id: AuthorId) {
+    fun canDelete(toDo: Todo): Boolean {
+        return toDo.authorId == id
     }
 }
